+++ include/semaphore.h
@@ -12,11 +12,11 @@
 /* Semaphore table entry */
 struct	sentry	{
 	byte	sstate;		/* Whether entry is S_FREE or S_USED	*/
-	int32	scount;		/* Count for the semaphore		*/
+	int16	scount;		/* Count for the semaphore		*/
 	qid16	squeue;		/* Queue of processes that are waiting	*/
-				/*     on the semaphore			*/
+						/*     on the semaphore			*/
 };
 
 extern	struct	sentry semtab[];
 
-#define	isbadsem(s)	((int32)(s) < 0 || (s) >= NSEM)
+#define	isbadsem(s)	((int16)(s) < 0 || (s) >= NSEM)
